workflows/pyproject.toml
[tool.poetry]
name = "legal-document-workflows"
version = "0.1.0"
description = "Temporal workflows for legal document processing"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
temporalio = "^1.5.0"
python-dotenv = "^1.0.0"
asyncio = "^3.4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
worker = "workflows.worker:main" 